refactor: move inside procedure folder
