+++ src/trpc/routers/stakeholder-router/procedures/remove-member.ts
@@ -1,21 +1,18 @@
 import { adminOnlyProcedure } from "@/trpc/api/trpc";
-import { ZodDeactivateUserMutationSchema } from "./schema";
+import { ZodRemoveMemberMutationSchema } from "../schema";
 import { Audit } from "@/server/audit";
 
-export const deactivateUserProcedure = adminOnlyProcedure
-  .input(ZodDeactivateUserMutationSchema)
+export const removeMemberProcedure = adminOnlyProcedure
+  .input(ZodRemoveMemberMutationSchema)
   .mutation(async ({ ctx: { session, db }, input }) => {
     const user = session.user;
-    const { membershipId, status } = input;
+    const { membershipId } = input;
 
     await db.$transaction(async (tx) => {
-      const member = await tx.membership.update({
+      const member = await tx.membership.delete({
         where: {
           id: membershipId,
           companyId: session.user.companyId,
-        },
-        data: {
-          active: status,
         },
         select: {
           userId: true,
@@ -34,14 +31,12 @@ export const deactivateUserProcedure = adminOnlyProcedure
 
       await Audit.create(
         {
-          action: status ? "stakeholder.activated" : "stakeholder.deactivated",
+          action: "stakeholder.removed",
           companyId: user.companyId,
           actor: { type: "user", id: user.id },
           context: {},
           target: [{ type: "user", id: member.userId }],
-          summary: `${user.name} ${
-            status ? "activated" : "deactivated"
-          } ${member.user?.name} from ${member?.company.name}`,
+          summary: `${user.name} removed ${member.user?.name} from ${member?.company?.name}`,
         },
         tx,
       );
