+++ src/trpc/routers/stakeholder-router/router.ts
@@ -1,13 +1,13 @@
 import { createTRPCRouter } from "@/trpc/api/trpc";
 
-import { inviteMemberProcedure } from "./invite-member";
-import { acceptMemberProcedure } from "./accept-member";
-import { revokeInviteProcedure } from "./revoke-invite";
-import { removeMemberProcedure } from "./remove-member";
-import { deactivateUserProcedure } from "./deactivate-user";
-import { updateMemberProcedure } from "./update-member";
-import { reInviteProcedure } from "./re-invite";
-import { getMembersProcedure } from "./get-members";
+import { inviteMemberProcedure } from "./procedures/invite-member";
+import { acceptMemberProcedure } from "./procedures/accept-member";
+import { revokeInviteProcedure } from "./procedures/revoke-invite";
+import { removeMemberProcedure } from "./procedures/remove-member";
+import { deactivateUserProcedure } from "./procedures/deactivate-user";
+import { updateMemberProcedure } from "./procedures/update-member";
+import { reInviteProcedure } from "./procedures/re-invite";
+import { getMembersProcedure } from "./procedures/get-members";
 
 export const stakeholderRouter = createTRPCRouter({
   inviteMember: inviteMemberProcedure,
